# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TradeMind Quant Engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/mcp/tools/get_market_regime": {
      "post": {
        "summary": "Get Market Regime",
        "operationId": "get_market_regime_mcp_tools_get_market_regime_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/calculate_var": {
      "post": {
        "summary": "Calculate Var",
        "operationId": "calculate_var_mcp_tools_calculate_var_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/check_graham_filters": {
      "post": {
        "summary": "Check Graham Filters",
        "operationId": "check_graham_filters_mcp_tools_check_graham_filters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/market_data_quality": {
      "post": {
        "summary": "Market Data Quality",
        "operationId": "market_data_quality_mcp_tools_market_data_quality_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MarketDataQualityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/historical_data_normalizer": {
      "post": {
        "summary": "Historical Data Normalizer",
        "operationId": "historical_data_normalizer_mcp_tools_historical_data_normalizer_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HistoricalDataNormalizerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/provider_fallback": {
      "post": {
        "summary": "Provider Fallback",
        "operationId": "provider_fallback_mcp_tools_provider_fallback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderFallbackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/signal_quality": {
      "post": {
        "summary": "Signal Quality",
        "operationId": "signal_quality_mcp_tools_signal_quality_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignalQualityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/robust_backtest": {
      "post": {
        "summary": "Robust Backtest",
        "operationId": "robust_backtest_mcp_tools_robust_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RobustBacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/portfolio_risk_manager": {
      "post": {
        "summary": "Portfolio Risk Manager",
        "operationId": "portfolio_risk_manager_mcp_tools_portfolio_risk_manager_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioRiskManagerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/mcp/tools/trade_execution_guard": {
      "post": {
        "summary": "Trade Execution Guard",
        "operationId": "trade_execution_guard_mcp_tools_trade_execution_guard_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TradeExecutionGuardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ml/extract_features": {
      "post": {
        "summary": "Extract Features",
        "operationId": "extract_features_ml_extract_features_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ml/predict_direction": {
      "post": {
        "summary": "Predict Direction",
        "operationId": "predict_direction_ml_predict_direction_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ml/predict_sarima": {
      "post": {
        "summary": "Predict Sarima",
        "operationId": "predict_sarima_ml_predict_sarima_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/lean/backtest": {
      "post": {
        "summary": "Lean Backtest",
        "operationId": "lean_backtest_lean_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/workflow/analyze": {
      "post": {
        "summary": "Run Workflow",
        "operationId": "run_workflow_workflow_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SymbolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ml/trigger_sentiment_scan": {
      "post": {
        "summary": "Trigger Sentiment Scan",
        "operationId": "trigger_sentiment_scan_ml_trigger_sentiment_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchSymbolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BatchSymbolRequest": {
        "properties": {
          "symbols": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Symbols"
          }
        },
        "type": "object",
        "required": [
          "symbols"
        ],
        "title": "BatchSymbolRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoricalDataNormalizerRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "US"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "1d"
          },
          "raw_dataset": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Raw Dataset"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "provider"
        ],
        "title": "HistoricalDataNormalizerRequest"
      },
      "LeanRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "LeanRequest"
      },
      "MarketDataQualityRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "provider": {
            "type": "string",
            "title": "Provider",
            "default": "yahoo-chart"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "1d"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "dataset": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Dataset"
          },
          "metadata": {
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "MarketDataQualityRequest"
      },
...
      "ProviderFallbackRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "market": {
            "type": "string",
            "title": "Market",
            "default": "US"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "1d"
          },
          "range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Range",
            "default": "2y"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "required_use": {
            "type": "string",
            "title": "Required Use",
            "default": "ml"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "ProviderFallbackRequest"
      },
...
      "SymbolRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "SymbolRequest"
      },
      "ToolRequest": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "timeframe": {
            "type": "string",
            "title": "Timeframe",
            "default": "1D"
          }
        },
        "type": "object",
        "required": [
          "symbol"
        ],
        "title": "ToolRequest"
      },
...
    }
  }
}